requirements.txt
-i https://pypi.org/simple
dj-database-url==0.5.0
django-cors-headers==3.2.1
django-heroku==0.3.1
django==2.2.7
djangorestframework==3.10.3
gunicorn==20.0.2
psycopg2==2.7.7
pytz==2019.3
sqlparse==0.3.0
whitenoise==5.0.1
